requirements.txt
selenium>=4.0.0
webdriver-manager>=3.8.0
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.0.0
